Remove calculated fields from Application PDF
Was running into error when I tried to combine just the Application and Order
into their own ALDocumentBundle. The error specifically is triggered when the
Application PDF is the first doc in a pdf_concatenate. When trying to get that
bundle, I get this:

```
Traceback (most recent call last):
  ...
  File "/usr/share/docassemble/local3.8/lib/python3.8/site-packages/docassemble/AssemblyLine/al_document.py", line 1127, in download_list_html
    doc.as_pdf(
  File "/usr/share/docassemble/local3.8/lib/python3.8/site-packages/docassemble/AssemblyLine/al_document.py", line 936, in as_pdf
    pdf = pdf_concatenate(
  File "/usr/share/docassemble/local3.8/lib/python3.8/site-packages/docassemble/base/util.py", line 8050, in pdf_concatenate
    pdf_path = docassemble.base.pandoc.concatenate_files(paths, pdfa=kwargs.get('pdfa', False), password=kwargs.get('password', None))
  File "/usr/share/docassemble/local3.8/lib/python3.8/site-packages/docassemble/base/pandoc.py", line 741, in concatenate_files
    replicate_js_and_calculations(new_path_list[0], pdf_file.name, password)
  File "/usr/share/docassemble/local3.8/lib/python3.8/site-packages/docassemble/base/pdftk.py", line 571, in replicate_js_and_calculations
    field = f.getObject()
  File "/usr/share/docassemble/local3.8/lib/python3.8/site-packages/PyPDF2/generic.py", line 177, in getObject
    return self.pdf.getObject(self).getObject()
  File "/usr/share/docassemble/local3.8/lib/python3.8/site-packages/PyPDF2/pdf.py", line 1706, in getObject
    raise utils.PdfReadError("Could not find object.")
PyPDF2.utils.PdfReadError: Could not find object.
```

Introspecting the object, there are some issues:

The code that's calling it in docassemble looks like this (failing on line 571)

```python
    template = safe_pypdf_reader(template_filename)
    co_field_names = []
    if '/AcroForm' in template.trailer['/Root']:
        #logmessage("Found AcroForm")
        acroform = template.trailer['/Root']['/AcroForm'].getObject()
        if '/CO' in acroform:
            #logmessage("Found CO in AcroForm")
            for f in acroform['/CO']:
                field = f.getObject()  # line 571
                if '/TM' in field:
                    name = field['/TM']
```

The 'CO' is the "array of indirect references to field dictionaries with
calculation actions, defining the calculation order in which their values will
be recalculated when the value of any field changes". So it's because the field
has a calculation, specific the sum of the employment, child support, etc.
fields. In Adobe, these seem to be the older names for the fields ("4B - Money
from Other Amount, 4B - OtherAmount, 4B - Child Support Amount, etc.). The
object itself really isn't there, but only after docassemble makes the
attachment file. Getting the individual ALDocument for the file works fine (and
was how I was able to introspect it).

The real fix is likely somewhere in the docassemble code that handles the
generating attachments, but the quicker fix is to remove calculations from
the PDF, which this commit does.